Split wasm_utils memory handling code into separate files.
