+++ core/src/hash_table/status/mod.rs
@@ -69,7 +69,6 @@ mod tests {
         eav::tests::eav_round_trip_test_runner,
         storage::{tests::ExampleContentAddressableStorage, ContentAddressableStorage},
     };
-    use hash::HashString;
 
     #[test]
     /// test the CrudStatus bit flags as ints
